requirements.txt
numpy==2.1.1
pyside6==6.7.2
python-snap7==2.0.0
QtPy==2.4.1
pyinstaller==6.10.0
opencv-python~=4.10.0.84